backend/requirements.txt
fastapi==0.109.0
httpx==0.26.0
passlib==1.7.4
pydantic==2.5.3
SQLAlchemy==1.4.51
uvicorn==0.25.0
Babel==2.14.0
black==23.12.1
httpx==0.26.0
mysql-connector-python==8.3.0
python-jose==3.3.0
python-multipart==0.0.5
pytest 
pytest-asyncio
pytest-mock
trio>=0.19.0
respx